Make DB calls async (#89)
* Make DB calls async

* Fixed DB errors (see description)

Error type 1:
No data persistence due to many contexts. Fixed by adding ComputerService as a transient service in Startup.

Error type 2:
Exceptions thrown when using the site because a new action was started on the context before the previous was finished. Fixed by using the same context and letting it finish when it's done - synchronous work.

* Removed unused code

* Add support for searching for missing assets + reset search on fail

* Good catch! Fixed.